feat: proj3-task3

+++ saniraq/main.py
@@ -1,7 +1,8 @@
-from fastapi import FastAPI, HTTPException, Depends, Response, Form
+from fastapi import FastAPI, HTTPException, Depends, Response, Form, status
+from fastapi.responses import JSONResponse
 from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.orm import Session
-from jose import jwt
+from jose import JWTError, jwt
 
 from .database import SessionLocal, Base, engine
 from .actions import UsersRepository, AdsRepository, CommentsRepository
@@ -83,17 +84,24 @@ def post_login(
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/users/login")
 
-def decode_jwt_token(token: str) -> int:
-    data = jwt.decode(token, "kek", "HS256")
-    return data["user_id"]
+def verify_token(token: str = Depends(oauth2_scheme)):
+    try:
+        data = jwt.decode(token, "kek", "HS256")
+        return data["user_id"]
+    except jwt.InvalidTokenError:
+        return HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            content={"message": "Invalid or expired token"},
+            headers={"WWW-Authenticate": "Bearer"},
+        )
 
 @app.patch("/auth/users/me")
 def patch_update_user(
     user_update: UserUpdate,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
+    
     updated_user = users_repository.update_user(db, user_id=current_user_id, new_info=user_update)
     
     if not updated_user:
@@ -107,9 +115,8 @@ def patch_update_user(
 @app.get("/auth/users/me")
 def get_user_info(
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     current_user_info = users_repository.get_by_id(db=db, user_id=current_user_id)
 
     return current_user_info
@@ -121,11 +128,9 @@ def get_user_info(
 def post_ad(
     input_ad: AdCreate,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     saved_ad = ads_repository.save_ad(db=db, ad=input_ad, user_id=current_user_id)
-    
     return {"ad_id": saved_ad.id}
 
 # ------------ TASK6 - GET AD ------
@@ -153,16 +158,14 @@ def patch_update_ad(
         id: int,
         ad_update: AdCreate,
         db: Session = Depends(get_db),
-        token: str = Depends(oauth2_scheme)
+        current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     found_ad = ads_repository.get_by_id(db=db, ad_id=id)
 
     if not found_ad:
         raise HTTPException(status_code=404, detail="Not found ad")
     
     ads_repository.update_ad(db=db, ad_id=found_ad.id, new_info=ad_update)
-
     return Response("Updated - OK", status_code=200)
 
 
@@ -172,9 +175,8 @@ def patch_update_ad(
 def delete_ad(
     id: int,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     found_ad = ads_repository.get_by_id(db, ad_id=id)
 
     if not found_ad:
@@ -195,16 +197,14 @@ def post_comments(
     id: int,
     new_comment: CommentCreate,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
     current_ad = ads_repository.get_by_id(db=db, ad_id=id)
-    current_user_id = decode_jwt_token(token)
 
     if not current_ad:
         raise HTTPException(status_code=404, detail="Not found ad")
 
     comments_repository.save_comment(db=db, comment=new_comment, user_id=current_user_id, ad_id=current_ad.id)
-
     return Response("Comment created - OK", status_code=200)
 
 
@@ -225,7 +225,7 @@ def get_comments(
     return {"comments": all_comment_by_ad_id}
 
 
-# ------------ TASK11 - UPDATE COMMENTS ------
+# ------------ TASK11 - UPDATE COMMENTS ------ +TASK3
 
 @app.patch("/shanyraks/{id}/comments/{comment_id}")
 def patch_update_comments(
@@ -233,36 +233,48 @@ def patch_update_comments(
     comment_id: int,
     new_info: CommentUpdate,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     current_ad = ads_repository.get_by_id(db=db, ad_id=id)
     current_comment = comments_repository.get_by_id(db=db, comment_id=comment_id)
 
     if not current_ad or not current_comment:
         raise HTTPException(status_code=404, detail="Not found")
     
-    comments_repository.update_comment(db=db, comment_id=current_comment.id, new_info=new_info)
+    current_ad_owner_id = current_ad.user_id
 
+    if current_ad_owner_id != current_user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to modify other users' comments"
+        )
+    
+    comments_repository.update_comment(db=db, comment_id=current_comment.id, new_info=new_info)
     return Response("Updated - OK", status_code=200)
 
 
-# ------------ TASK12 - DELETE COMMENTS ------
+# ------------ TASK12 - DELETE COMMENTS ------ +TASK3
 
 @app.delete("/shanyraks/{id}/comments/{comment_id}")
 def delete_comment(
     id: int,
     comment_id: int,
     db: Session = Depends(get_db),
-    token: str = Depends(oauth2_scheme)
+    current_user_id: int = Depends(verify_token)
 ):
-    current_user_id = decode_jwt_token(token)
     current_ad = ads_repository.get_by_id(db=db, ad_id=id)
     current_comment = comments_repository.get_by_id(db=db, comment_id=comment_id)
 
     if not current_ad or not current_comment:
         raise HTTPException(status_code=404, detail="Not found")
     
-    comments_repository.delete_comment(db=db, comment_id=current_comment.id)
+    current_ad_owner_id = current_ad.user_id
 
+    if current_ad_owner_id != current_user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to delete other users' comments"
+        )
+    
+    comments_repository.delete_comment(db=db, comment_id=current_comment.id)
     return Response("Deleted - OK", status_code=200)
